+++ integration_tests/iceberg/test_iceberg.py
@@ -2,17 +2,17 @@ import json
 import os
 import pickle
 import threading
-import time
 import uuid
 
 import pandas as pd
 import pytest
 from dateutil import tz
 from pyiceberg.catalog import load_catalog
+from pyiceberg.catalog.glue import GlueCatalog
 
 import pathway as pw
 from pathway.internals.parse_graph import G
-from pathway.tests.utils import run, wait_result_with_checker
+from pathway.tests.utils import get_aws_s3_settings, run, wait_result_with_checker
 
 INPUT_CONTENTS_1 = """{"user_id": 1, "name": "John"}
 {"user_id": 2, "name": "Jane"}
@@ -27,7 +27,16 @@ INPUT_CONTENTS_3 = """{"user_id": 9, "name": "Anna"}
 {"user_id": 11, "name": "Steve"}
 {"user_id": 12, "name": "Sarah"}"""
 
-CATALOG_URI = os.environ.get("ICEBERG_CATALOG_URI", "http://iceberg:8181")
+LOCAL_BACKEND_NAME = "local"
+S3_BACKEND_NAME = "s3"
+CATALOG_URI = {
+    LOCAL_BACKEND_NAME: os.environ.get("ICEBERG_CATALOG_URI", "http://iceberg:8181"),
+    S3_BACKEND_NAME: os.environ.get("ICEBERG_S3_CATALOG_URI", "http://iceberg-s3:8181"),
+}
+S3_CONNECTION_SETTINGS = {
+    LOCAL_BACKEND_NAME: None,
+    S3_BACKEND_NAME: get_aws_s3_settings(),
+}
 INPUT_CONTENTS = {
     1: INPUT_CONTENTS_1,
     2: INPUT_CONTENTS_2,
@@ -35,31 +44,51 @@ INPUT_CONTENTS = {
 }
 
 
-def _get_pandas_table(table_name: str) -> pd.DataFrame:
-    catalog = load_catalog(name="default", uri=CATALOG_URI)
+def _get_pandas_table(backend: str, table_name: str) -> pd.DataFrame:
+    if backend == LOCAL_BACKEND_NAME:
+        catalog = load_catalog(name="default", uri=CATALOG_URI[backend])
+    elif backend == S3_BACKEND_NAME:
+        catalog = GlueCatalog(name="default")
+    else:
+        raise RuntimeError(f"Unknown backend type: {backend}")
     table = catalog.load_table(table_name)
     scan = table.scan()
     return scan.to_pandas()
 
 
+def _get_warehouse_path(backend: str, table_name: str) -> str | None:
+    if backend == LOCAL_BACKEND_NAME:
+        return None
+    return f"s3://aws-integrationtest/iceberg/{table_name}"
+
+
 class IcebergEntriesCountChecker:
-    def __init__(self, table_name: str, expected_count: int):
+    def __init__(self, backend: str, table_name: str, expected_count: int):
+        self.backend = backend
         self.table_name = table_name
         self.expected_count = expected_count
 
-    def __call__(self):
+    def __call__(self) -> bool:
         try:
-            table = _get_pandas_table(self.table_name)
+            table = _get_pandas_table(self.backend, self.table_name)
             return len(table) == self.expected_count
         except Exception:
             return False
 
+    def provide_information_on_failure(self) -> str:
+        try:
+            table = _get_pandas_table(self.backend, self.table_name)
+            return f"Table length is: {len(table)}. Expected count is {self.expected_count}"
+        except Exception as e:
+            return f"Failed to read the table. Exception: {e}"
 
-def test_iceberg_read_after_write(tmp_path):
+
+@pytest.mark.parametrize("backend", [LOCAL_BACKEND_NAME, S3_BACKEND_NAME])
+def test_iceberg_read_after_write(backend, tmp_path):
     input_path = tmp_path / "input.txt"
     output_path = tmp_path / "output.txt"
     pstorage_path = tmp_path / "pstorage"
-    table_name = str(uuid.uuid4())
+    table_name = str(uuid.uuid4()).replace("-", "")
 
     def run_single_iteration(seq_number: int):
         input_path.write_text(INPUT_CONTENTS[seq_number])
@@ -80,20 +109,24 @@ def test_iceberg_read_after_write(tmp_path):
         )
         pw.io.iceberg.write(
             table,
-            catalog_uri=CATALOG_URI,
+            catalog_uri=CATALOG_URI[backend],
             namespace=["my_database"],
             table_name=table_name,
+            warehouse=_get_warehouse_path(backend, table_name),
+            s3_connection_settings=S3_CONNECTION_SETTINGS[backend],
         )
         run()
 
         # Read the data from the Iceberg table via Pathway
         G.clear()
         table = pw.io.iceberg.read(
-            catalog_uri=CATALOG_URI,
+            catalog_uri=CATALOG_URI[backend],
             namespace=["my_database"],
             table_name=table_name,
             mode="static",
             schema=InputSchema,
+            warehouse=_get_warehouse_path(backend, table_name),
+            s3_connection_settings=S3_CONNECTION_SETTINGS[backend],
         )
         pw.io.jsonlines.write(table, output_path)
         persistence_config = pw.persistence.Config(
@@ -113,9 +146,10 @@ def test_iceberg_read_after_write(tmp_path):
 
 
 @pytest.mark.flaky(reruns=5)
-def test_iceberg_several_runs(tmp_path):
+@pytest.mark.parametrize("backend", [LOCAL_BACKEND_NAME, S3_BACKEND_NAME])
+def test_iceberg_several_runs(backend, tmp_path):
     input_path = tmp_path / "input.txt"
-    table_name = str(uuid.uuid4())
+    table_name = str(uuid.uuid4()).replace("-", "")
     all_ids = set()
     all_names = set()
 
@@ -138,14 +172,16 @@ def test_iceberg_several_runs(tmp_path):
         )
         pw.io.iceberg.write(
             table,
-            catalog_uri=CATALOG_URI,
+            catalog_uri=CATALOG_URI[backend],
             namespace=["my_database"],
             table_name=table_name,
+            warehouse=_get_warehouse_path(backend, table_name),
+            s3_connection_settings=S3_CONNECTION_SETTINGS[backend],
         )
         run()
 
         iceberg_table_name = f"my_database.{table_name}"
-        pandas_table = _get_pandas_table(iceberg_table_name)
+        pandas_table = _get_pandas_table(backend, iceberg_table_name)
         assert pandas_table.shape == (4 * seq_number, 4)
         assert set(pandas_table["user_id"]) == all_ids
         assert set(pandas_table["name"]) == all_names
@@ -158,16 +194,22 @@ def test_iceberg_several_runs(tmp_path):
         run_single_iteration(seq_number)
 
 
-def test_iceberg_streaming(tmp_path):
+@pytest.mark.parametrize("backend", ["local", "s3"])
+def test_iceberg_streaming(backend, tmp_path):
     inputs_path = tmp_path / "inputs"
     inputs_path.mkdir()
-    table_name = str(uuid.uuid4())
+    table_name = str(uuid.uuid4()).replace("-", "")
 
     def stream_inputs():
         for i in range(1, len(INPUT_CONTENTS) + 1):
             input_path = inputs_path / f"{i}.txt"
             input_path.write_text(INPUT_CONTENTS[i])
-            time.sleep(5)
+            wait_result_with_checker(
+                IcebergEntriesCountChecker(backend, f"my_database.{table_name}", 4 * i),
+                30,
+                step=1,
+                target=None,
+            )
 
     class InputSchema(pw.Schema):
         user_id: int
@@ -180,17 +222,19 @@ def test_iceberg_streaming(tmp_path):
     )
     pw.io.iceberg.write(
         table,
-        catalog_uri=CATALOG_URI,
+        catalog_uri=CATALOG_URI[backend],
         namespace=["my_database"],
         table_name=table_name,
         min_commit_frequency=1000,
+        warehouse=_get_warehouse_path(backend, table_name),
+        s3_connection_settings=S3_CONNECTION_SETTINGS[backend],
     )
 
     t = threading.Thread(target=stream_inputs)
     t.start()
     wait_result_with_checker(
         IcebergEntriesCountChecker(
-            f"my_database.{table_name}", 4 * len(INPUT_CONTENTS)
+            backend, f"my_database.{table_name}", 4 * len(INPUT_CONTENTS)
         ),
         30,
     )
@@ -204,7 +248,7 @@ def test_iceberg_streaming(tmp_path):
             all_names.add(data["name"])
 
     iceberg_table_name = f"my_database.{table_name}"
-    pandas_table = _get_pandas_table(iceberg_table_name)
+    pandas_table = _get_pandas_table(backend, iceberg_table_name)
     assert pandas_table.shape == (4 * len(INPUT_CONTENTS), 4)
     assert set(pandas_table["user_id"]) == all_ids
     assert set(pandas_table["name"]) == all_names
@@ -212,10 +256,11 @@ def test_iceberg_streaming(tmp_path):
     assert len(set(pandas_table["time"])) == len(INPUT_CONTENTS)
 
 
-def test_py_object_wrapper_in_iceberg(tmp_path):
+@pytest.mark.parametrize("backend", ["local", "s3"])
+def test_py_object_wrapper_in_iceberg(backend, tmp_path):
     input_path = tmp_path / "input.jsonl"
     output_path = tmp_path / "output.jsonl"
-    iceberg_table_name = str(uuid.uuid4())
+    iceberg_table_name = str(uuid.uuid4()).replace("-", "")
     input_path.write_text("test")
 
     table = pw.io.plaintext.read(input_path, mode="static")
@@ -225,9 +270,11 @@ def test_py_object_wrapper_in_iceberg(tmp_path):
     )
     pw.io.iceberg.write(
         table,
-        catalog_uri=CATALOG_URI,
+        catalog_uri=CATALOG_URI[backend],
         namespace=["my_database"],
         table_name=iceberg_table_name,
+        warehouse=_get_warehouse_path(backend, iceberg_table_name),
+        s3_connection_settings=S3_CONNECTION_SETTINGS[backend],
     )
     run()
     G.clear()
@@ -241,11 +288,13 @@ def test_py_object_wrapper_in_iceberg(tmp_path):
         return a.value(x)
 
     table = pw.io.iceberg.read(
-        catalog_uri=CATALOG_URI,
+        catalog_uri=CATALOG_URI[backend],
         namespace=["my_database"],
         table_name=iceberg_table_name,
         mode="static",
         schema=InputSchema,
+        warehouse=_get_warehouse_path(backend, iceberg_table_name),
+        s3_connection_settings=S3_CONNECTION_SETTINGS[backend],
     )
     table = table.select(len=use_python_object(pw.this.fun, pw.this.data))
     pw.io.jsonlines.write(table, output_path)
@@ -256,9 +305,10 @@ def test_py_object_wrapper_in_iceberg(tmp_path):
         assert data["len"] == 4
 
 
-def test_iceberg_different_types_serialization(tmp_path):
+@pytest.mark.parametrize("backend", ["local", "s3"])
+def test_iceberg_different_types_serialization(backend, tmp_path):
     input_path = tmp_path / "input.jsonl"
-    iceberg_table_name = str(uuid.uuid4())
+    iceberg_table_name = str(uuid.uuid4()).replace("-", "")
     input_path.write_text("test")
 
     column_values = {
@@ -279,9 +329,11 @@ def test_iceberg_different_types_serialization(tmp_path):
     )
     pw.io.iceberg.write(
         table,
-        catalog_uri=CATALOG_URI,
+        catalog_uri=CATALOG_URI[backend],
         namespace=["my_database"],
         table_name=iceberg_table_name,
+        warehouse=_get_warehouse_path(backend, iceberg_table_name),
+        s3_connection_settings=S3_CONNECTION_SETTINGS[backend],
     )
     run()
     G.clear()
@@ -308,11 +360,13 @@ def test_iceberg_different_types_serialization(tmp_path):
                 assert row[field] == expected_value
 
     table = pw.io.iceberg.read(
-        catalog_uri=CATALOG_URI,
+        catalog_uri=CATALOG_URI[backend],
         namespace=["my_database"],
         table_name=iceberg_table_name,
         mode="static",
         schema=InputSchema,
+        warehouse=_get_warehouse_path(backend, iceberg_table_name),
+        s3_connection_settings=S3_CONNECTION_SETTINGS[backend],
     )
     checker = Checker()
     pw.io.subscribe(table, on_change=checker)